requirement.txt
langchain-google-genai
google-generativeai
gradio
langchain 
python-dotenv